move test to a new file and clean up storage object
